requirements.txt
python-telegram-bot==21.4
SQLAlchemy==2.0.0
requests==2.32.3
